Inline FoodsMixin methods into Fatsecret class
